User commit

+++ src/main/java/com/example/demo/repository/validationrepository.java
@@ -1,9 +1,9 @@
 package com.example.demo.repository;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.example.demo.entity.studententity;
+import com.example.demo.entity.validationentity;
 
 @Repository
-public interface validationrepository extends JpaRepository<validationentity,Integer>{
+public interface validationrepository extends JpaRepository<validationentity,Long>{
 
 }
